% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwnnet.R
\name{predict.rwnnet}
\alias{predict.rwnnet}
\title{Predict Method for rwnnet}
\usage{
\method{predict}{rwnnet}(object, x, ...)
}
\arguments{
\item{object}{a fitted object of class inherithng from "rwnnet".}

\item{x}{data.frame or tibble in which to look for inputs with which to
predict. If omitted, the leave-one-out cross-validation results are used.}

\item{...}{...}
}
\description{
Predict Method for rwnnet
}
